+++ app/src/main/java/org/groover/bar/data/item/ItemRepository.kt
@@ -12,7 +12,7 @@ class ItemRepository(
     Item.Companion::deserialize,
     listOf("ID", "Name", "Price", "BTW Percentage"),
 ) {
-    fun addItem(name: String, price: Float, btwPercentage: Int) {
+    fun addItem(name: String, price: Int, btwPercentage: Int) {
         // Create new item
         val newItem = Item(
             generateId(),
@@ -30,7 +30,7 @@ class ItemRepository(
 
 
 
-    fun changeItem(itemId: Int, newName: String, newPrice: Float, newBtwPercentage: Int) {
+    fun changeItem(itemId: Int, newName: String, newPrice: Int, newBtwPercentage: Int) {
         // Create new item
         val newItem = Item(
             itemId,
